Upating OpenCV and Pygame examples
