+++ opencv/detect.py
@@ -28,6 +28,7 @@ python3 detect.py \
 """
 import argparse
 import collections
+import common
 import cv2
 import numpy as np
 import os
@@ -35,7 +36,6 @@ from PIL import Image
 import re
 import tflite_runtime.interpreter as tflite
 
-EDGETPU_SHARED_LIB = 'libedgetpu.so.1'
 Object = collections.namedtuple('Object', ['id', 'score', 'bbox'])
 
 def load_labels(path):
@@ -43,39 +43,6 @@ def load_labels(path):
     with open(path, 'r', encoding='utf-8') as f:
        lines = (p.match(line).groups() for line in f.readlines())
        return {int(num): text.strip() for num, text in lines}
-
-def make_interpreter(model_file):
-    model_file, *device = model_file.split('@')
-    return tflite.Interpreter(
-      model_path=model_file,
-      experimental_delegates=[
-          tflite.load_delegate(EDGETPU_SHARED_LIB,
-                               {'device': device[0]} if device else {})
-      ])
-
-def input_image_size(interpreter):
-    """Returns input image size as (width, height) tuple."""
-    _, height, width, _ = interpreter.get_input_details()[0]['shape']
-    return width, height
-
-def input_tensor(interpreter):
-    """Returns input tensor view as numpy array of shape (height, width, 3)."""
-    tensor_index = interpreter.get_input_details()[0]['index']
-    return interpreter.tensor(tensor_index)()[0]
-
-def output_tensor(interpreter, i):
-    """Returns output tensor view."""
-    tensor = interpreter.tensor(interpreter.get_output_details()[i]['index'])()
-    return np.squeeze(tensor)
-
-def set_input(interpreter, data):
-    """Copies data to input tensor."""
-    input_tensor(interpreter)[:, :] = data
-
-def set_interpreter(interpreter, image, resample=Image.NEAREST):
-    image = image.resize((input_image_size(interpreter)), resample)
-    set_input(interpreter, image)
-    interpreter.invoke()
 
 class BBox(collections.namedtuple('BBox', ['xmin', 'ymin', 'xmax', 'ymax'])):
     """Bounding box.
@@ -86,10 +53,10 @@ class BBox(collections.namedtuple('BBox', ['xmin', 'ymin', 'xmax', 'ymax'])):
 
 def get_output(interpreter, score_threshold, top_k, image_scale=1.0):
     """Returns list of detected objects."""
-    boxes = output_tensor(interpreter, 0)
-    class_ids = output_tensor(interpreter, 1)
-    scores = output_tensor(interpreter, 2)
-    count = int(output_tensor(interpreter, 3))
+    boxes = common.output_tensor(interpreter, 0)
+    class_ids = common.output_tensor(interpreter, 1)
+    scores = common.output_tensor(interpreter, 2)
+    count = int(common.output_tensor(interpreter, 3))
 
     def make(i):
         ymin, xmin, ymax, xmax = boxes[i]
@@ -119,7 +86,7 @@ def main():
     args = parser.parse_args()
 
     print('Loading {} with {} labels.'.format(args.model, args.labels))
-    interpreter = make_interpreter(args.model)
+    interpreter = common.make_interpreter(args.model)
     interpreter.allocate_tensors()
     labels = load_labels(args.labels)
 
@@ -133,7 +100,8 @@ def main():
 
         pil_im = Image.fromarray(cv2_im)
 
-        set_interpreter(interpreter, pil_im)
+        common.set_input(interpreter, pil_im)
+        interpreter.invoke()
         objs = get_output(interpreter, score_threshold=args.threshold, top_k=args.top_k)
         cv2_im = append_objs_to_img(cv2_im, objs, labels)
 
@@ -143,7 +111,6 @@ def main():
 
     cap.release()
     cv2.destroyAllWindows()
-
 
 def append_objs_to_img(cv2_im, objs, labels):
     height, width, channels = cv2_im.shape
@@ -157,7 +124,6 @@ def append_objs_to_img(cv2_im, objs, labels):
         cv2_im = cv2.putText(cv2_im, label, (x0, y0+30),
                              cv2.FONT_HERSHEY_SIMPLEX, 1.0, (255, 0, 0), 2)
     return cv2_im
-
 
 if __name__ == '__main__':
     main()
