+++ pygame/classify_capture.py
@@ -16,6 +16,7 @@
 import argparse
 import collections
 from collections import deque
+import common
 import io
 import numpy as np
 import operator
@@ -27,42 +28,15 @@ import tflite_runtime.interpreter as tflite
 import time
 
 Category = collections.namedtuple('Category', ['id', 'score'])
-EDGETPU_SHARED_LIB = 'libedgetpu.so.1'
-
-def make_interpreter(model_file):
-    model_file, *device = model_file.split('@')
-    return tflite.Interpreter(
-      model_path=model_file,
-      experimental_delegates=[
-          tflite.load_delegate(EDGETPU_SHARED_LIB,
-                               {'device': device[0]} if device else {})
-      ])
-
-def input_image_size(interpreter):
-    """Returns input image size as (width, height, channels) tuple."""
-    _, height, width, channels = interpreter.get_input_details()[0]['shape']
-    return width, height, channels
 
 def input_tensor(interpreter):
     """Returns input tensor view as numpy array of shape (height, width, 3)."""
     tensor_index = interpreter.get_input_details()[0]['index']
     return interpreter.tensor(tensor_index)()[0]
 
-def output_tensor(interpreter):
-    """Returns dequantized output tensor."""
-    output_details = interpreter.get_output_details()[0]
-    output_data = np.squeeze(interpreter.tensor(output_details['index'])())
-    scale, zero_point = output_details['quantization']
-    return scale * (output_data - zero_point)
-
-def set_interpreter(interpreter, data):
-    """Copies data to input tensor."""
-    input_tensor(interpreter)[:,:] = np.reshape(data, (input_image_size(interpreter)))
-    interpreter.invoke()
-
 def get_output(interpreter, top_k, score_threshold):
     """Returns no more than top_k categories with score >= score_threshold."""
-    scores = output_tensor(interpreter)
+    scores = common.output_tensor(interpreter, 0)
     categories = [
         Category(i, scores[i])
         for i in np.argpartition(scores, -top_k)[-top_k:]
@@ -85,7 +59,7 @@ def main():
         pairs = (l.strip().split(maxsplit=1) for l in f.readlines())
         labels = dict((int(k), v) for k, v in pairs)
 
-    interpreter = make_interpreter(args.model)
+    interpreter = common.make_interpreter(args.model)
     interpreter.allocate_tensors()
 
     pygame.init()
@@ -94,7 +68,7 @@ def main():
 
     print('By default using camera: ', camlist[-1])
     camera = pygame.camera.Camera(camlist[-1], (640, 480)) 
-    width, height, channels = input_size(interpreter)
+    width, height, channels = common.input_image_size(interpreter)
     camera.start()
     try:
         fps = deque(maxlen=20)
@@ -104,7 +78,8 @@ def main():
             imagen = pygame.transform.scale(imagen, (width, height))
             input = np.frombuffer(imagen.get_buffer(), dtype=np.uint8)
             start_ms = time.time()
-            set_interpreter(interpreter, input)
+            common.input_tensor(interpreter)[:,:] = np.reshape(input, (common.input_image_size(interpreter)))
+            interpreter.invoke()
             results = get_output(interpreter, top_k=3, score_threshold=0)
             inference_ms = (time.time() - start_ms)*1000.0
             fps.append(time.time())
