+++ pygame/detect.py
@@ -16,6 +16,7 @@
 import argparse
 import collections
 from collections import deque
+import common
 import io
 import numpy as np
 import os
@@ -26,7 +27,6 @@ import re
 import tflite_runtime.interpreter as tflite
 import time
 
-EDGETPU_SHARED_LIB = 'libedgetpu.so.1'
 Object = collections.namedtuple('Object', ['id', 'score', 'bbox'])
 
 def load_labels(path):
@@ -35,32 +35,13 @@ def load_labels(path):
        lines = (p.match(line).groups() for line in f.readlines())
        return {int(num): text.strip() for num, text in lines}
 
-def make_interpreter(model_file):
-    model_file, *device = model_file.split('@')
-    return tflite.Interpreter(
-      model_path=model_file,
-      experimental_delegates=[
-          tflite.load_delegate(EDGETPU_SHARED_LIB,
-                               {'device': device[0]} if device else {})
-      ])
-
-def input_image_size(interpreter):
-    """Returns input image size as (width, height, channels) 3-tuple."""
-    _, height, width, channels = interpreter.get_input_details()[0]['shape']
-    return width, height, channels
-
 def input_tensor(interpreter):
     """Returns input tensor view as numpy array of shape (height, width, 3)."""
     tensor_index = interpreter.get_input_details()[0]['index']
     return interpreter.tensor(tensor_index)()[0]
 
-def output_tensor(interpreter, i):
-    """Returns output tensor view."""
-    tensor = interpreter.tensor(interpreter.get_output_details()[i]['index'])()
-    return np.squeeze(tensor)
-
 def set_interpreter(interpreter, data):
-    input_tensor(interpreter)[:,:] = np.reshape(data, (input_image_size(interpreter)))
+    input_tensor(interpreter)[:,:] = np.reshape(data, (common.input_image_size(interpreter)))
     interpreter.invoke()
 
 class BBox(collections.namedtuple('BBox', ['xmin', 'ymin', 'xmax', 'ymax'])):
@@ -72,9 +53,9 @@ class BBox(collections.namedtuple('BBox', ['xmin', 'ymin', 'xmax', 'ymax'])):
 
 def get_output(interpreter, score_threshold, top_k, image_scale=1.0):
     """Returns list of detected objects."""
-    boxes = output_tensor(interpreter, 0)
-    class_ids = output_tensor(interpreter, 1)
-    scores = output_tensor(interpreter, 2)
+    boxes = common.output_tensor(interpreter, 0)
+    class_ids = common.output_tensor(interpreter, 1)
+    scores = common.output_tensor(interpreter, 2)
 
     def make(i):
         ymin, xmin, ymax, xmax = boxes[i]
@@ -110,7 +91,7 @@ def main():
 
     print('Loading {} with {} labels.'.format(args.model, args.labels))
 
-    interpreter = make_interpreter(args.model)
+    interpreter = common.make_interpreter(args.model)
     interpreter.allocate_tensors()
     labels = load_labels(args.labels)
 
@@ -121,8 +102,8 @@ def main():
     pygame.camera.init()
     camlist = pygame.camera.list_cameras()
 
-    w, h, _ = input_image_size(interpreter)
-  
+    w, h, _ = common.input_image_size(interpreter)
+
     print('By default using camera: ', camlist[-1])
     camera = pygame.camera.Camera(camlist[-1], (cam_w, cam_h))
     try:
@@ -131,7 +112,7 @@ def main():
       sys.stderr.write("\nERROR: Unable to open a display window. Make sure a monitor is attached and that "
             "the DISPLAY environment variable is set. Example: \n"
             ">export DISPLAY=\":0\" \n")
-      raise e 
+      raise e
 
     red = pygame.Color(255, 0, 0)
 
@@ -143,7 +124,8 @@ def main():
             imagen = pygame.transform.scale(mysurface, (w, h))
             input = np.frombuffer(imagen.get_buffer(), dtype=np.uint8)
             start_time = time.monotonic()
-            set_interpreter(interpreter, input)
+            common.input_tensor(interpreter)[:,:] = np.reshape(input, (common.input_image_size(interpreter)))
+            interpreter.invoke()
             results = get_output(interpreter, score_threshold=args.threshold, top_k=args.top_k)
             stop_time = time.monotonic()
             inference_ms = (stop_time - start_time)*1000.0
